% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital_quiz.R
\name{captial_quiz}
\alias{captial_quiz}
\title{Capital quiz}
\usage{
captial_quiz(regions = c("all", "northeast", "southeast", "midwest",
  "southwest", "west"), type = c("fill", "mc"))
}
\arguments{
\item{regions}{a list of brms regions to use for quiz.  Included in
`states_caps`` data frame.}

\item{type}{Which type of test to take, can be "mc" of "fill".  Only "fill"
is currently implemented.}
}
\description{
This function gives a quiz for each of the capitals based on region or regions
}
